+++ test/resources/unused_l10n_analyzer/unused_l10n_analyzer_example.dart
@@ -11,6 +11,8 @@ void main() {
 
   TestI18n.of('').anotherRegularGetter;
 
+  TestI18n.of('').baseGetter;
+
   final wrapper = L10nWrapper();
 
   wrapper.l10n.regularField.trim();