+++ test/src/analyzers/unused_l10n_analyzer/unused_l10n_analyzer_test.dart
@@ -31,19 +31,17 @@ void main() {
           config,
         );
 
-        final report = result.single;
+        final report = result.firstWhere((report) => report.issues.length == 3);
         expect(report.className, 'TestI18n');
-
-        expect(report.issues.length, 3);
 
         final firstIssue = report.issues.first;
         expect(firstIssue.memberName, 'getter');
-        expect(firstIssue.location.line, 4);
+        expect(firstIssue.location.line, 6);
         expect(firstIssue.location.column, 3);
 
         final secondIssue = report.issues.elementAt(1);
         expect(secondIssue.memberName, 'regularGetter');
-        expect(secondIssue.location.line, 13);
+        expect(secondIssue.location.line, 15);
         expect(secondIssue.location.column, 3);
 
         final thirdIssue = report.issues.elementAt(2);
@@ -51,8 +49,17 @@ void main() {
           thirdIssue.memberName,
           'secondMethod(String value, num number)',
         );
-        expect(thirdIssue.location.line, 8);
+        expect(thirdIssue.location.line, 10);
         expect(thirdIssue.location.column, 3);
+
+        final parentReport =
+            result.firstWhere((report) => report.issues.length == 1);
+        expect(parentReport.className, 'TestI18n');
+
+        final firstParentIssue = parentReport.issues.first;
+        expect(firstParentIssue.memberName, 'anotherBaseGetter');
+        expect(firstParentIssue.location.line, 4);
+        expect(firstParentIssue.location.column, 3);
       });
 
       test('should analyze files with custom class pattern', () async {
@@ -74,17 +81,17 @@ void main() {
 
         final firstIssue = report.issues.first;
         expect(firstIssue.memberName, 'field');
-        expect(firstIssue.location.line, 25);
+        expect(firstIssue.location.line, 27);
         expect(firstIssue.location.column, 3);
 
         final secondIssue = report.issues.elementAt(1);
         expect(secondIssue.memberName, 'regularField');
-        expect(secondIssue.location.line, 42);
+        expect(secondIssue.location.line, 44);
         expect(secondIssue.location.column, 3);
 
         final thirdIssue = report.issues.elementAt(2);
         expect(thirdIssue.memberName, 'method(String value)');
-        expect(thirdIssue.location.line, 29);
+        expect(thirdIssue.location.line, 31);
         expect(thirdIssue.location.column, 3);
 
         final forthIssue = report.issues.elementAt(3);
@@ -92,7 +99,7 @@ void main() {
           forthIssue.memberName,
           'secondMethod(String value, num number)',
         );
-        expect(forthIssue.location.line, 31);
+        expect(forthIssue.location.line, 33);
         expect(forthIssue.location.column, 3);
       });