+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * chore: restrict `analyzer` version to `>=2.4.0 <3.3.0`.
+* fix: make `check-unused-l10n` also cover supertype member calls.
 
 ## 4.10.0-dev.2